refactor!: remove unused helper methods
